backend/requirements_dev.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
openai>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
pydantic[email]>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
python-dotenv>=1.0.0
jinja2>=3.1.2

# Enhanced UI/UX and Export Dependencies
reportlab>=4.0.4
openpyxl>=3.1.2
xlsxwriter>=3.1.2
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.15.0
kaleido>=0.2.1
email-validator>=2.0.0 